fix(celldb): round-1 review — rates stripped and ENFORCED, drift test, port occupancy
The 3/3 major was the schema's own headline invariant violated by its own
seed data: cloned entities carried fixture-specific rate stamps into
data/celldb.json. Rates are now stripped at extraction AND check_entry
refuses any entity storing one — the invariant is enforced, not narrated.
Likewise motif.count is re-derived from the fragment and asserted
(mis-ranking silently was the 1/3 major), and every port must sit on a
transport tile carrying its declared item.

Extraction moved into celldb::extract_unit (warnings returned, not
printed) so the new drift test rebuilds a seed source fixture with the
CURRENT engine and diffs motif/metrics/ports/roster against the stored
entry — stored-vs-stored testing cannot see engine drift (2/3 minor). The
seed tool is now a thin driver that counts escape hatches in its output
(K67-1 trips above 1; current: 0). is_splitter replaces the substring
poke.

Acknowledged, not fixed: sub-row demands always get the whole-row entry
(P1 scope, reviewer agrees); PipeOut exists in the schema but nothing
emits it — fluid-OUT rows are a recorded gap alongside community port
inference.
